[v4] WIP graduate PanelStack2
