+++ packages/core/src/components/panel-stack/_panel-stack2.scss
