+++ packages/core/src/components/panel-stack/panel-stack2.md
