+++ packages/core/src/components/panel-stack/panelStack2.tsx
