+++ packages/core/src/components/panel-stack/panelView2.tsx
